Fixing comments.
