--- accountmanager.go
@@ -1,35 +1,0 @@
-// Copyright 2022 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"net/http"
-)
-
-type UserInfo interface {
-	Username() string
-}
-
-type AuthHTTPHandler func(http.ResponseWriter, *http.Request, UserInfo) error
-
-type AccountManager interface {
-	// Returns the received http handler wrapped in another that extracts user
-	// information from the request and passes it to to the original handler as
-	// the last parameter.
-	// The wrapper will only pass the request to the inner handler if a user is
-	// authenticated, otherwise it may choose to return an error or respond with
-	// an HTTP redirect to the login page.
-	Authenticate(fn AuthHTTPHandler) HTTPHandler
-}
